requirements.txt
asgiref==3.5.0
Babel==2.9.1
CacheControl==0.12.10
certifi==2021.10.8
charset-normalizer==2.0.12
deprecation==2.1.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-money==2.1.1
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
jsonpickle==1.5.2
Markdown==3.3.6
msgpack==1.0.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pgeocode==0.3.0
Pillow==9.1.0
psycopg2==2.9.3
py-moneyed==1.2
pygeodesic==0.1.7
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
squareup==17.3.0.20220316
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
